push to render.com

+++ db/migrate/20241204134015_remove_address_from_users.rb
@@ -1,5 +1,7 @@
 class RemoveAddressFromUsers < ActiveRecord::Migration[6.1]
   def change
-    remove_column :users, :address, :string
+    if column_exists?(:users, :address)
+      remove_column :users, :address, :string
+    end
   end
 end
